Better code style following MVVM pattern
